Move specimen pivot queries to study-api and specimen.xsd to specimen module
