+++ study/api-src/org/labkey/api/specimen/SpecimenQuerySchema.java
@@ -1,11 +1,16 @@
 package org.labkey.api.specimen;
 
+import org.jetbrains.annotations.Nullable;
+import org.labkey.api.data.ContainerFilter;
+import org.labkey.api.data.TableInfo;
 import org.labkey.api.query.UserSchema;
 import org.labkey.api.security.User;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
 
-public class SpecimenQuerySchema
+import java.util.Set;
+
+public class SpecimenQuerySchema extends UserSchema
 {
     public static final String SIMPLE_SPECIMEN_TABLE_NAME = "SimpleSpecimen";
     public static final String SPECIMEN_DETAIL_TABLE_NAME = "SpecimenDetail";
@@ -20,8 +25,35 @@ public class SpecimenQuerySchema
     public static final String SPECIMEN_ADDITIVE_TABLE_NAME = "SpecimenAdditive";
     public static final String VIAL_TABLE_NAME = "Vial";
 
-    public static UserSchema get(Study study, User user)
+    private final UserSchema _studySchema;
+    private final Study _study;
+
+    public SpecimenQuerySchema(Study study, UserSchema studySchema)
     {
-        return StudyService.get().getStudyQuerySchema(study, user);
+        super(studySchema.getName(), studySchema.getDescription(), studySchema.getUser(), studySchema.getContainer(), studySchema.getDbSchema());
+        _study = study;
+        _studySchema = studySchema;
+    }
+
+    public static SpecimenQuerySchema get(Study study, User user)
+    {
+        return new SpecimenQuerySchema(study, StudyService.get().getStudyQuerySchema(study, user));
+    }
+
+    public Study getStudy()
+    {
+        return _study;
+    }
+
+    @Override
+    public @Nullable TableInfo createTable(String name, ContainerFilter cf)
+    {
+        return _studySchema.createTable(name, cf);
+    }
+
+    @Override
+    public Set<String> getTableNames()
+    {
+        return _studySchema.getTableNames();
     }
 }
